[FE] feat: signUp, dupilicationCheck slice 분리
